ref: move tester and pipeline to the pkg root
