--- checker/tester/__init__.py
@@ -1,1 +1,0 @@
-from .tester import Tester  # noqa: F401
